+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - **FEAT**: Add `ShadIconButton` component.
 - **REFACTOR**: Deprecated `ShadButton.icon`, use `ShadIconButton` for a button with just an icon and `ShadButton.leading` for a button with an icon and a text.
 - **REFACTOR**: Deprecate `orderPolicy`, use `leading` and `trailing` in the component.
+- **REFACTOR**: Deprecate `prefix` and `suffix` in `ShadInput`, use `leading` and `trailing` instead.
 
 ## 0.19.3